# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API for tracking usage and time spent on various web and desktop applications. Created with Express, Node.js and documentation by Swagger. Accessible via web front-end or REST API endpoints.",
        "version": "1.0.0",
        "title": "playtimetracker.app",
        "contact": {
            "email": "mail@rhysbi.shop"
        }
    },
    "host": "api.playtimetracker.app",
    "basePath": "/",
    "tags": [
        {
            "name": "github",
            "description": "Retrieve GitHub events produced by a user",
            "externalDocs": {
                "description": "Find out more",
                "url": "https://playtimetracker.app/"
            }
        },
        {
            "name": "steam",
            "description": "Retrieve recent playtime for a user"
        }
    ],
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
        "/github/{username}/{duration}": {
            "get": {
                "tags": [
                    "github"
                ],
                "summary": "Retrieve count of GitHub events produced by a user",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The user to retrieve the events of ",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "duration",
                        "in": "path",
                        "description": "The period in seconds to retrieve events from ",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/GitHub"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username or duration"
                    }
                }
            }
        },
        "/github/{username}/{duration}/{type}": {
            "get": {
                "tags": [
                    "github"
                ],
                "summary": "Retrieve count of GitHub events of a certain type produced by a user",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The user to retrieve the events of ",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "duration",
                        "in": "path",
                        "description": "The period in seconds to retrieve events from ",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "The type of event to retrieve ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/GitHubType"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username or duration"
                    }
                }
            }
        },
        "/steam/{id}": {
            "get": {
                "tags": [
                    "steam"
                ],
                "summary": "Retrieve recent playtime ",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "The user to retrieve the events of ",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "duration",
                        "in": "path",
                        "description": "The period in seconds to retrieve events from ",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type",
                        "in": "path",
                        "description": "The type of event to retrieve ",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/GitHubType"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid username or duration"
                    }
                }
            }
        }
    },
    "definitions": {
        "GitHub": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer",
                    "format": "int64"
                },
                "PushEvent": {
                    "type": "integer",
                    "format": "int64"
                },
                "WatchEvent": {
                    "type": "integer",
                    "format": "int64"
                },
                "CreateEvent": {
                    "type": "integer",
                    "format": "int64"
                },
                "PublicEvent": {
                    "type": "integer",
                    "format": "int64"
                },
                "ForkEvent": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "GitHubType": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    },
    "externalDocs": {
        "description": "Go back home",
        "url": "https://playtimetracker.app/"
    }
}